refactor: relocate focused helpers and audit checkpoints
