+++ docs/dev/audits/2026-06-26-agent-sdk-readiness-checkpoint.md
@@ -1,8 +1,8 @@
 # Agent SDK Readiness — Verification Checkpoint (2026-06-26)
 
 **Status:** Verification checkpoint, not a new audit. Read alongside the canonical
-[`2026-05-30-agent-sdk-readiness.md`](./2026-05-30-agent-sdk-readiness.md), the
-[`2026-06-24-agent-sdk-readiness-delta.md`](./2026-06-24-agent-sdk-readiness-delta.md)
+[`2026-05-30-agent-sdk-readiness.md`](../../proposals/2026-05-30-agent-sdk-readiness.md), the
+[`2026-06-24-agent-sdk-readiness-delta.md`](../../proposals/2026-06-24-agent-sdk-readiness-delta.md)
 addendum, and the [`2026-06-25-agent-sdk-readiness-checkpoint.md`](./2026-06-25-agent-sdk-readiness-checkpoint.md)
 checkpoint. This pass re-verified the standing audit against the working tree at
 HEAD (`93af483`) and records **only** what is genuinely new since the 2026-06-25
