+++ docs/dev/audits/2026-08-01-agent-sdk-readiness-checkpoint.md
@@ -4,10 +4,10 @@
 [`2026-07-30-agent-sdk-readiness-checkpoint.md`](./2026-07-30-agent-sdk-readiness-checkpoint.md)
 (this pass reconciles against its §New-8…§New-11 and [TRACKED] items rather than
 re-deriving them), the foundation-gap analysis
-[`2026-07-26-agent-sdk-foundation-gap.md`](./2026-07-26-agent-sdk-foundation-gap.md)
+[`2026-07-26-agent-sdk-foundation-gap.md`](../../proposals/2026-07-26-agent-sdk-foundation-gap.md)
 (§9 the real ceiling), the audit of record
-[`../dev/audits/2026-07-25-agent-sdk-readiness-audit.md`](../dev/audits/2026-07-25-agent-sdk-readiness-audit.md),
-the plan of record [`2026-07-09-agent-sdk-north-star.md`](./2026-07-09-agent-sdk-north-star.md)
+[`../dev/audits/2026-07-25-agent-sdk-readiness-audit.md`](./2026-07-25-agent-sdk-readiness-audit.md),
+the plan of record [`2026-07-09-agent-sdk-north-star.md`](../../proposals/2026-07-09-agent-sdk-north-star.md)
 (§4 sequenced path, §5 verified traps), and the `-06-25` → `-07-30` checkpoint chain.
 
 This pass inspected the tree afresh at HEAD `6ab67ce`
