+++ src/agent/modelHandlers/utils/toolAttachmentUtils.ts
@@ -7,7 +7,7 @@ import {
 // Local imports - utils
 import { isNonEmptyString } from '@utils/core';
 import { WorkspaceFS } from '@utils/files';
-import { appendHead, appendTail } from '@utils/strings/appendTail';
+import { appendHead, appendTail } from '@utils/text/appendTail';
 
 // Local imports - model handlers
 import {
