+++ src/test-kernel/agent/runtime/RetryState.vitest.ts
@@ -56,7 +56,7 @@ import { DEFAULT_CORE_SETTINGS } from '@shared/schemas/coreSettings';
 import {
   clearStreamStatusForTest,
   seedStreamStatusForTest,
-} from '@test/helpers/streamStatusTestUtils';
+} from '@test/support/streamStatusTestUtils';
 import { createTestSession } from '@test/support/sessionTestUtils';
 
 // Local file imports
