+++ src/test-kernel/support/streamStatusTestUtils.ts
