+++ src/utils/text/appendTail.ts
